requirements.txt
openpyxl>=3.1.0
requests>=2.31.0
Pillow>=10.0.0
